Finished code commenting, added pointer use, quick README setup, added data to gitignore, fixed indentation, intuitively named variables, removed of config
