+++ .gitignore
@@ -1,0 +1,6 @@
+# ignore all other file in cmd directory except .go files
+cmd/*/*
+!cmd/*/*.go
+*/cmd/*/*
+!*/cmd/*/*.go
+assets/*.json
