--- app.env
@@ -1,3 +1,0 @@
-OUT_DIR=./assets/news_site_articles.json
-CRYPTO_BUZZ='Altcoin', 'crypto', 'bitcoin', 'btc', 'ethereum', 'eth', 'staking', 'proof of concept', 'smart contract', 'market', 'stock', 'binance', 'ftx', 'coinbase', 'kucoin', 'robinhood', 'ripple', 'lawsuit', 'decentralized'
-WAR_POLITICS_BUZZ='attack on', 'people killed', 'troops advancing', 'russia', 'america', 'election', 'protest', 'banks', 'bankrupt', 'ukraine'
