--- main.go
@@ -1,11 +1,0 @@
-/*
-Copyright © 2022 four-cryptics
-*/
-
-package main
-
-import "github.com/four-cryptics/crypto-news-notifier/cmd"
-
-func main() {
-	cmd.Execute()
-}
